Minor QoL update and removal of scrollbar
Customized draggable marker image, increased size

Moved font-family from globals to MapUI.module.css

Removed globals.css, something seems to be overriding it and applying an 8px margin on the page's body and display: none; on scrollbar was not working...

Due to above issue, the only solution to remove scrollbars from the mapbox was to add -8px margin to Mapbox.module.css. Any attempt to hide scrollbars in other ways did not work, this is temporary anyway.